backend/pyproject.toml
[project]
name = "rag-handbook-backend"
version = "0.1.0"
description = "FastAPI backend for RAG Company Handbook"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "google-generativeai>=0.8.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-chroma>=0.1.0",
    "langchain-community>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "chromadb>=0.5.0",
    "openai>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
